% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averaged_cram.R
\name{averaged_cram}
\alias{averaged_cram}
\title{Averaged CRAM with Permutations}
\usage{
averaged_cram(
  X,
  D,
  Y,
  batch,
  model_type,
  learner_type = NULL,
  alpha = 0.05,
  baseline_policy = NULL,
  parallelize_batch = FALSE,
  model_params = NULL,
  custom_fit = NULL,
  custom_predict = NULL,
  num_permutations = 10
)
}
\arguments{
\item{X}{A matrix or data frame of covariates.}

\item{D}{A binary vector of treatment indicators (0 or 1).}

\item{Y}{A vector of outcomes.}

\item{batch}{Either an integer specifying the number of batches (which will be created by random sampling) or a vector of length equal to the sample size providing the batch assignment (index) for each individual in the sample.}

\item{model_type}{The model type for policy learning. Options include \code{"causal_forest"}, \code{"s_learner"}, and \code{"m_learner"}. Default is \code{"causal_forest"}.}

\item{learner_type}{The learner type for the chosen model. Options include \code{"ridge"} for Ridge Regression and \code{"fnn"} for Feedforward Neural Network. Default is \code{"ridge"}.}

\item{alpha}{Significance level for confidence intervals. Default is 0.05 (95\% confidence).}

\item{baseline_policy}{A list providing the baseline policy (binary 0 or 1) for each sample. If \code{NULL}, defaults to a list of zeros with the same length as the number of rows in \code{X}.}

\item{parallelize_batch}{Logical. Whether to parallelize batch processing (i.e. the cram method learns T policies, with T the number of batches. They are learned in parallel when parallelize_batch is TRUE vs. learned sequentially using the efficient data.table structure when parallelize_batch is FALSE, recommended for light weight training). Defaults to \code{FALSE}.}

\item{model_params}{A list of additional parameters to pass to the model, which can be any parameter defined in the model reference package. Defaults to \code{NULL}.}

\item{custom_fit}{A custom, user-defined, function that outputs a fitted model given training data (allows flexibility). Defaults to \code{NULL}.}

\item{custom_predict}{A custom, user-defined, function for making predictions given a fitted model and test data (allow flexibility). Defaults to \code{NULL}.}

\item{num_permutations}{Number of random permutations of batches.}
}
\value{
A list with averaged performance and variance estimates.
}
\description{
This function implements Averaged CRAM by randomly permuting batches
(except the last batch which is kept the same) and averaging performance results.
}
\examples{
X <- matrix(rnorm(1000), nrow = 100, ncol = 10)
D <- sample(0:1, 100, replace = TRUE)
Y <- rnorm(100)
avg_cram_results <- averaged_cram(X, D, Y,
                                  batch = 20,
                                  model_type = "m_learner",
                                  learner_type = "ridge",
                                  num_permutations = 3)
}
