% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_bandit.R
\name{cram_bandit}
\alias{cram_bandit}
\title{CRAM Bandit - On-policy Statistical Evaluation in Contextual Bandit}
\usage{
cram_bandit(pi, arm, reward, batch = 1, alpha = 0.05)
}
\arguments{
\item{pi}{for each row j, column t, depth a, gives pi_t(Xj, a)}

\item{arm}{arms selected at each time step}

\item{reward}{rewards at each time step}

\item{batch}{Batch size}

\item{alpha}{Confidence level for intervals (default = 0.05).}
}
\value{
A list containing:
\itemize{
  \item{raw_results: Data frame with performance metrics}
  \item{interactive_table: DT::datatable interactive view}
  \item{final_ml_model: Trained model object}
}
}
\description{
Performs the Cram method for On-policy Statistical Evaluation
in Contextual Bandit
}
\seealso{
\code{\link[caret]{train}} for model training parameters
\code{\link[stats]{kmeans}} for unsupervised clustering
\code{\link[DT]{datatable}} for interactive tables
}
