% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armed_bandit_var.R
\name{cram_bandit_var}
\alias{cram_bandit_var}
\title{Cram Bandit Variance}
\usage{
cram_bandit_var(pi, reward, arm, batch = 1)
}
\arguments{
\item{pi}{A 3-d array, for each row j, column t, depth a, gives pi_t(Xj, a)}

\item{reward}{A vector of rewards}

\item{arm}{A vector of arms chosen}

\item{batch}{Batch size}
}
\value{
The crammed variance estimate for bandit.
}
\description{
This function implements the crammed variance estimator for bandit policy
}
