% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_estimate.R
\name{cram_estimator}
\alias{cram_estimator}
\title{Cram Estimator for Policy Evaluation}
\usage{
cram_estimator(Y, D, pi, batch_indices)
}
\arguments{
\item{Y}{A vector of outcomes for the n individuals.}

\item{D}{A vector of binary treatments for the n individuals.}

\item{pi}{A matrix of n rows and (nb_batch + 1) columns, containing the policy assignment for each individual
for each policy. The first column represents the baseline policy.}

\item{batch_indices}{A list where each element is a vector of indices corresponding to the individuals in each batch.}
}
\value{
The estimated policy value difference \(\eqn{\hat{\Delta}}(\eqn{\hat{\pi}}_T; \eqn{\pi}_0)\).
}
\description{
This function estimates the difference in policy performance between a series of learned policies
and a baseline policy using the cram approach.
}
\examples{
# Example usage:
Y <- sample(0:1, 100, replace = TRUE)
D <- sample(0:1, 100, replace = TRUE)
pi <- matrix(sample(0:1, 100 * 11, replace = TRUE), nrow = 100, ncol = 11)
nb_batch <- 10
batch_indices <- split(1:100, rep(1:nb_batch, each = 10))
estimate <- cram_estimator(Y, D, pi, batch_indices)
}
