% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armed_bandit_var.R
\name{crammed_variance_estimator}
\alias{crammed_variance_estimator}
\title{Crammed Variance Estimator}
\usage{
crammed_variance_estimator(Y, D, pi, X, T)
}
\arguments{
\item{Y}{A vector of observed rewards for each time step.}

\item{D}{A vector of treatment indicators (1 if treated, 0 if not) for each time step.}

\item{pi}{A matrix where each row represents the policy probabilities for actions at time t.}

\item{X}{A matrix or data frame of covariates observed at each time step.}

\item{T}{The total number of iterations in the bandit process.}
}
\value{
The crammed variance estimate \eqn{\hat{\sigma}^2_T}.
}
\description{
This function implements the crammed variance estimator \eqn{\hat{\sigma}^2_T}
as described in the provided formula.
}
