% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict_ml.R
\name{model_predict_ml}
\alias{model_predict_ml}
\title{Predict with the Specified Model}
\usage{
model_predict_ml(model, data, formula, caret_params)
}
\arguments{
\item{model}{A trained model object returned by the `fit_model_ml` function.}

\item{data}{The dataset}

\item{formula}{The formula}

\item{caret_params}{The parameters of the caret model}
}
\value{
A vector of predictions or CATE estimates, depending on the \code{model_type} and \code{learner_type}.
}
\description{
This function performs inference using a trained model
}
\examples{
# Load required library
library(grf)

# Example: Predicting with a Causal Forest model
set.seed(123)
X <- matrix(rnorm(1000), nrow = 100, ncol = 10)  # Covariates
Y <- rnorm(100)                                 # Outcomes
D <- sample(0:1, 100, replace = TRUE)           # Treatment indicators
cf_model <- causal_forest(X, Y, D)             # Train Causal Forest
new_X <- matrix(rnorm(100), nrow = 10, ncol = 10) # New data for predictions
predictions <- model_predict(model = cf_model, X = new_X, model_type = "causal_forest")
}
\seealso{
\code{\link[grf]{causal_forest}}, \code{\link[glmnet]{cv.glmnet}}
}
