% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armed_bandit_est.R
\name{cram_bandit_est}
\alias{cram_bandit_est}
\title{Cramming Policy Evaluation for Multi-Armed Bandit}
\usage{
cram_bandit_est(pi, reward, arm, batch = 1)
}
\arguments{
\item{pi}{A 3-d array, for each row j, column t, depth a, gives pi_t(Xj, a)}

\item{reward}{A vector of rewards}

\item{arm}{A vector of arms chosen}

\item{batch}{A vector or integer. If a vector, gives the batch assignment for each context.
If an integer, interpreted as the batch size and contexts are assigned to a batch in the order of the dataset.}
}
\value{
The estimated policy value difference \eqn{\Delta(\pi_T; \pi_0)}.
}
\description{
This function implements the armed bandit policy evaluation formula for
estimating \eqn{\Delta(\pi_T; \pi_0)} as given in the user-provided formula.
}
