% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_bandit_sim.R
\name{cram_bandit_sim}
\alias{cram_bandit_sim}
\title{On-policy CRAM Bandit}
\usage{
cram_bandit_sim(
  horizon,
  simulations,
  bandit,
  policy,
  alpha = 0.05,
  do_parallel = FALSE,
  seed = 42
)
}
\arguments{
\item{horizon}{The number of timesteps}

\item{simulations}{The number of simulations}

\item{bandit}{The bandit, generating contextual features and observed rewards}

\item{policy}{The policy, choosing the arm at each timestep}

\item{alpha}{Significance level for confidence intervals for calculating the empirical coverage. Default is 0.05 (95\% confidence).}

\item{do_parallel}{Whether to parallelize the simulations. Default to FALSE.}

\item{seed}{Seed for simulation}
}
\value{
A **list** containing:
  \item{raw_results}{A data frame summarizing key metrics:
  Average Prediction Error on the Policy Value Estimate,
  Average Prediction Error on the Variance Estimate,
  Empirical Coverage of the alpha level Confidence Interval.}
  \item{interactive_table}{An interactive table summarizing key metrics for detailed exploration.}
}
\description{
This function performs simulation for cram bandit policy learning and evaluation.
}
