% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_var_expected_loss.R
\name{cram_var_expected_loss}
\alias{cram_var_expected_loss}
\title{Cram ML: Variance Estimate of the crammed expected loss estimate}
\usage{
cram_var_expected_loss(loss, batch_indices)
}
\arguments{
\item{loss}{A matrix of loss values with N rows (data points) and K+1 columns (batches).
We assume that the first column of the loss matrix contains only zeros.
The following nb_batch columns contain the losses of each trained model for each individual.}

\item{batch_indices}{A list where each element is a vector of indices corresponding to a batch.}
}
\value{
The variance estimate of the crammed expected loss estimate
}
\description{
This function computes the variance estimator based on the given loss matrix and batch indices.
}
