% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict_ml.R
\name{model_predict_ml}
\alias{model_predict_ml}
\title{Predict with the Specified Model}
\usage{
model_predict_ml(model, data, formula, caret_params)
}
\arguments{
\item{model}{A trained model object returned by the `fit_model_ml` function.}

\item{data}{The dataset}

\item{formula}{The formula}

\item{caret_params}{The parameters of the caret model}
}
\value{
A vector of predictions or CATE estimates, depending on the \code{model_type} and \code{learner_type}.
}
\description{
This function performs inference using a trained model
}
\seealso{
\code{\link[grf]{causal_forest}}, \code{\link[glmnet]{cv.glmnet}}
}
