% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict_ml.R
\name{model_predict_ml}
\alias{model_predict_ml}
\title{Cram ML: Predict with the Specified Model}
\usage{
model_predict_ml(
  model,
  data,
  formula,
  caret_params,
  cram_policy_handle = FALSE
)
}
\arguments{
\item{model}{A trained model object returned by the `fit_model_ml` function.}

\item{data}{The dataset}

\item{formula}{The formula}

\item{caret_params}{The parameters of the caret model}

\item{cram_policy_handle}{Internal use. Post-process predictions differently for cram policy use. Defaults to FALSE.}
}
\value{
Predictions of the model on the data
}
\description{
This function performs inference using a trained model
}
