% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cram_helpers.R
\name{test_batch}
\alias{test_batch}
\title{Validate or Generate Batch Assignments}
\usage{
test_batch(batch, n)
}
\arguments{
\item{batch}{Either an integer specifying the number of batches or a vector/list of batch assignments for all individuals.}

\item{n}{An integer specifying the number of individuals in the population.}
}
\value{
A list containing:
  \describe{
    \item{\code{batches}}{A list where each element contains the indices of individuals assigned to a specific batch.}
    \item{\code{nb_batch}}{The total number of batches.}
  }
}
\description{
This function validates a provided batch assignment or generates random batch assignments for individuals.
}
