% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_params_fnn.R
\name{validate_params_fnn}
\alias{validate_params_fnn}
\title{Validate Parameters for Feedforward Neural Networks (FNNs)}
\usage{
validate_params_fnn(model_type, learner_type, model_params, X)
}
\arguments{
\item{model_type}{The type of model to validate. Supported options are \code{"causal_forest"},
\code{"s_learner"}, and \code{"m_learner"}.}

\item{learner_type}{The type of learner for the model. Options depend on \code{model_type}.
For example, \code{"ridge"} and \code{"fnn"} are supported learners.}

\item{model_params}{A named list of parameters provided by the user for configuring the FNN model.}

\item{X}{A matrix or data frame of covariates for which the parameters are validated.}
}
\value{
A named list of validated parameters merged with defaults for any missing values.
}
\description{
This function validates user-provided parameters for a Feedforward Neural Network (FNN) model.
It ensures the correct structure for \code{input_layer}, \code{layers}, \code{output_layer}, and \code{compile_args}.
}
