% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_params_fnn.R
\name{validate_params_fnn}
\alias{validate_params_fnn}
\title{Validate Parameters for Feedforward Neural Networks (FNNs)}
\usage{
validate_params_fnn(model_type, learner_type, model_params)
}
\arguments{
\item{model_params}{A named list of parameters provided by the user for configuring the FNN model.}
}
\value{
A named list of validated parameters merged with defaults for any missing values.
}
\description{
This function validates user-provided parameters for a Feedforward Neural Network (FNN) model.
It ensures the correct structure for \code{input_layer}, \code{layers}, \code{output_layer}, and \code{compile_args}.
}
